中小公司没有那么多数据量，Int64的自增改回来Int32，以降低复杂度

+++ AntJob.Data/Entity/作业任务.cs
@@ -18,13 +18,13 @@ namespace AntJob.Data.Entity
     public partial class JobTask : IJobTask
     {
         #region 属性
-        private Int64 _ID;
+        private Int32 _ID;
         /// <summary>编号</summary>
         [DisplayName("编号")]
         [Description("编号")]
         [DataObjectField(true, true, false, 0)]
         [BindColumn("ID", "编号", "")]
-        public Int64 ID { get { return _ID; } set { if (OnPropertyChanging(__.ID, value)) { _ID = value; OnPropertyChanged(__.ID); } } }
+        public Int32 ID { get { return _ID; } set { if (OnPropertyChanging(__.ID, value)) { _ID = value; OnPropertyChanged(__.ID); } } }
 
         private Int32 _AppID;
         /// <summary>应用</summary>
@@ -243,7 +243,7 @@ namespace AntJob.Data.Entity
             {
                 switch (name)
                 {
-                    case __.ID : _ID = value.ToLong(); break;
+                    case __.ID : _ID = value.ToInt(); break;
                     case __.AppID : _AppID = value.ToInt(); break;
                     case __.JobID : _JobID = value.ToInt(); break;
                     case __.Client : _Client = Convert.ToString(value); break;
@@ -428,7 +428,7 @@ namespace AntJob.Data.Entity
     {
         #region 属性
         /// <summary>编号</summary>
-        Int64 ID { get; set; }
+        Int32 ID { get; set; }
 
         /// <summary>应用</summary>
         Int32 AppID { get; set; }
